pyproject.toml
[tool.poetry]
name = "parking-app-backend"
version = "0.1.0"
description = ""
authors = ["т"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.105.0"
uvicorn = "0.24.0.post1"
sqlalchemy = "2.0.23"
pydantic = {extras = ["email"], version = "^2.6.4"}
requests = "2.31.0"
bcrypt = "4.0.1"
python-jose = "3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
